# app/main.py
from __future__ import annotations

import hmac
from contextlib import asynccontextmanager
from typing import Annotated

from fastapi import Depends, FastAPI, Header, HTTPException, Query
from fastapi.responses import PlainTextResponse
from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
from pydantic import BaseModel

from app import bootstrap as bootstrap_service
from app import sessions as session_service
from app import turns as turn_service
from app.config import ACTION_TOKEN, ENVIRONMENT, ROOT_DIR, ensure_data_dirs
from app.models import (
    AbortTurnRequest,
    AbortTurnResponse,
    BootstrapPartRequest,
    BootstrapSaveResponse,
    BootstrapValidationResponse,
    CommitTurnRequest,
    CreateSessionRequest,
    LegacyCreateSessionRequest,
    PrepareTurnRequest,
    PrepareTurnResponse,
    QuestionnaireRequest,
    SessionSummary,
    StartQuestionnaireResponse,
    TurnReceipt,
)


class HealthResponse(BaseModel):
    status: str
    service: str
    environment: str


bearer = HTTPBearer(auto_error=False)


def require_action_token(
    credentials: Annotated[HTTPAuthorizationCredentials | None, Depends(bearer)],
    x_api_key: Annotated[str | None, Header(alias="X-API-Key")] = None,
) -> None:
    if not ACTION_TOKEN:
        return
    bearer_token = (
        credentials.credentials
        if credentials is not None and credentials.scheme.lower() == "bearer"
        else ""
    )
    if not (
        hmac.compare_digest(bearer_token, ACTION_TOKEN)
        or hmac.compare_digest(x_api_key or "", ACTION_TOKEN)
    ):
        raise HTTPException(status_code=401, detail="Invalid action token")


@asynccontextmanager
async def lifespan(_: FastAPI):
    ensure_data_dirs()
    yield


app = FastAPI(
    title="Novel Runtime API",
    version="1.3.0",
    description="Persistent session runtime for a Custom GPT novella generator.",
    docs_url=None if ENVIRONMENT == "production" else "/docs",
    redoc_url=None if ENVIRONMENT == "production" else "/redoc",
    lifespan=lifespan,
)


@app.get("/health", response_model=HealthResponse, tags=["system"])
def health() -> HealthResponse:
    ensure_data_dirs()
    return HealthResponse(
        status="ok",
        service="novel-runtime",
        environment=ENVIRONMENT,
    )


@app.get("/openapi-actions.yaml", response_class=PlainTextResponse, include_in_schema=False)
def action_schema() -> PlainTextResponse:
    content = (ROOT_DIR / "openapi-actions.yaml").read_text(encoding="utf-8")
    return PlainTextResponse(content=content, media_type="application/yaml")


auth = [Depends(require_action_token)]


@app.get(
    "/api/v1/start-questionnaire",
    response_model=StartQuestionnaireResponse,
    dependencies=auth,
    tags=["bootstrap"],
    operation_id="getStartQuestionnaire",
)
@app.get(
    "/v1/start-questionnaire",
    response_model=StartQuestionnaireResponse,
    dependencies=auth,
    tags=["bootstrap"],
    include_in_schema=False,
)
def get_start_questionnaire() -> StartQuestionnaireResponse:
    questionnaire = (ROOT_DIR / "gpt" / "start_questionnaire.md").read_text(
        encoding="utf-8"
    )
    return StartQuestionnaireResponse(version="2.0", questionnaire=questionnaire)


@app.post(
    "/api/v1/sessions",
    response_model=SessionSummary,
    dependencies=auth,
    tags=["sessions"],
    include_in_schema=False,
)
def create_legacy_session(request: LegacyCreateSessionRequest) -> SessionSummary:
    return session_service.create_session(request)


@app.post("/v1/sessions", response_model=SessionSummary, dependencies=auth, tags=["sessions"])
def create_session(request: CreateSessionRequest) -> SessionSummary:
    return session_service.create_session(request)


@app.get(
    "/api/v1/sessions",
    response_model=list[SessionSummary],
    dependencies=auth,
    tags=["sessions"],
    include_in_schema=False,
)
@app.get("/v1/sessions", response_model=list[SessionSummary], dependencies=auth, tags=["sessions"])
def list_sessions(limit: int = Query(default=20, ge=1, le=50)) -> list[SessionSummary]:
    if not ACTION_TOKEN:
        raise HTTPException(
            status_code=403,
            detail="Session listing is disabled without a shared action secret; use resumeSession",
        )
    return session_service.list_sessions(limit)


@app.get(
    "/api/v1/sessions/resume/{resume_code}",
    response_model=SessionSummary,
    dependencies=auth,
    tags=["sessions"],
    include_in_schema=False,
)
@app.get(
    "/v1/sessions/resume/{resume_code}",
    response_model=SessionSummary,
    dependencies=auth,
    tags=["sessions"],
)
def resume_session(resume_code: str) -> SessionSummary:
    return session_service.resume_session(resume_code)


@app.get(
    "/api/v1/sessions/{session_id}",
    response_model=SessionSummary,
    dependencies=auth,
    tags=["sessions"],
    include_in_schema=False,
)
@app.get(
    "/v1/sessions/{session_id}",
    response_model=SessionSummary,
    dependencies=auth,
    tags=["sessions"],
)
def get_session_status(session_id: str) -> SessionSummary:
    return session_service.get_session_summary(session_id)


@app.put(
    "/api/v1/sessions/{session_id}/questionnaire",
    response_model=SessionSummary,
    dependencies=auth,
    tags=["bootstrap"],
    include_in_schema=False,
)
@app.put(
    "/v1/sessions/{session_id}/questionnaire",
    response_model=SessionSummary,
    dependencies=auth,
    tags=["bootstrap"],
)
def save_questionnaire(session_id: str, request: QuestionnaireRequest) -> SessionSummary:
    return bootstrap_service.save_questionnaire(session_id, request)


@app.post(
    "/api/v1/sessions/{session_id}/bootstrap/parts",
    response_model=BootstrapSaveResponse,
    dependencies=auth,
    tags=["bootstrap"],
    include_in_schema=False,
)
@app.post(
    "/v1/sessions/{session_id}/bootstrap/parts",
    response_model=BootstrapSaveResponse,
    dependencies=auth,
    tags=["bootstrap"],
)
def save_bootstrap_part(
    session_id: str,
    request: BootstrapPartRequest,
) -> BootstrapSaveResponse:
    return bootstrap_service.save_bootstrap_part(session_id, request)


@app.post(
    "/api/v1/sessions/{session_id}/bootstrap/validate",
    response_model=BootstrapValidationResponse,
    dependencies=auth,
    tags=["bootstrap"],
    include_in_schema=False,
)
@app.post(
    "/v1/sessions/{session_id}/bootstrap/validate",
    response_model=BootstrapValidationResponse,
    dependencies=auth,
    tags=["bootstrap"],
)
def validate_bootstrap(session_id: str) -> BootstrapValidationResponse:
    return bootstrap_service.validate_session_bootstrap(session_id)


@app.post(
    "/api/v1/sessions/{session_id}/bootstrap/confirm",
    response_model=SessionSummary,
    dependencies=auth,
    tags=["bootstrap"],
    include_in_schema=False,
)
@app.post(
    "/v1/sessions/{session_id}/bootstrap/confirm",
    response_model=SessionSummary,
    dependencies=auth,
    tags=["bootstrap"],
)
def confirm_bootstrap(session_id: str) -> SessionSummary:
    return bootstrap_service.confirm_bootstrap(session_id)


@app.post(
    "/api/v1/sessions/{session_id}/turns",
    response_model=PrepareTurnResponse,
    dependencies=auth,
    tags=["turns"],
    include_in_schema=False,
)
@app.post(
    "/v1/sessions/{session_id}/turns",
    response_model=PrepareTurnResponse,
    dependencies=auth,
    tags=["turns"],
)
def prepare_turn(session_id: str, request: PrepareTurnRequest) -> PrepareTurnResponse:
    return turn_service.prepare_turn(session_id, request)


@app.get(
    "/api/v1/sessions/{session_id}/turns/{turn_id}/chunks/{chunk_index}",
    response_model=PrepareTurnResponse,
    dependencies=auth,
    tags=["turns"],
    include_in_schema=False,
)
@app.get(
    "/v1/sessions/{session_id}/turns/{turn_id}/chunks/{chunk_index}",
    response_model=PrepareTurnResponse,
    dependencies=auth,
    tags=["turns"],
)
def get_turn_chunk(
    session_id: str,
    turn_id: str,
    chunk_index: int,
) -> PrepareTurnResponse:
    return turn_service.get_turn_chunk(session_id, turn_id, chunk_index)


@app.post(
    "/api/v1/sessions/{session_id}/turns/{turn_id}/commit",
    response_model=TurnReceipt,
    dependencies=auth,
    tags=["turns"],
    include_in_schema=False,
)
@app.post(
    "/v1/sessions/{session_id}/turns/{turn_id}/commit",
    response_model=TurnReceipt,
    dependencies=auth,
    tags=["turns"],
)
def commit_turn(
    session_id: str,
    turn_id: str,
    request: CommitTurnRequest,
) -> TurnReceipt:
    return turn_service.commit_turn(session_id, turn_id, request)


@app.post(
    "/api/v1/sessions/{session_id}/turns/{turn_id}/abort",
    response_model=AbortTurnResponse,
    dependencies=auth,
    tags=["turns"],
    include_in_schema=False,
)
@app.post(
    "/v1/sessions/{session_id}/turns/{turn_id}/abort",
    response_model=AbortTurnResponse,
    dependencies=auth,
    tags=["turns"],
)
def abort_turn(
    session_id: str,
    turn_id: str,
    request: AbortTurnRequest,
) -> AbortTurnResponse:
    return turn_service.abort_turn(session_id, turn_id, request.reason)