# openapi-actions.yaml
openapi: 3.1.0
info:
  title: Novel Runtime Actions
  version: 1.3.0
  description: Persistent file-backed sessions for a Custom GPT novella director.
servers:
  - url: https://web-production-4310e.up.railway.app
paths:
  /api/v1/start-questionnaire:
    get:
      operationId: getStartQuestionnaire
      summary: Load the visible starter questionnaire
      description: >-
        Returns the exact questionnaire to show when the user asks to start
        without giving story preferences. Show it verbatim, then wait for the
        answer before creating a session. It does not read or mix session state.
      x-openai-isConsequential: false
      responses:
        "200":
          description: Visible starter questionnaire
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartQuestionnaireResponse"
  /v1/sessions:
    post:
      operationId: createSession
      summary: Create a novel session with its questionnaire
      description: >-
        Creates one isolated session and saves all visible starter answers in
        the same operation. Call only after preferences or “Рандом” are visible.
        Supply raw_answers from the conversation; never ask the user to repeat it.
      x-openai-isConsequential: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "200":
          description: New session status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionSummary"
    get:
      operationId: listSessions
      summary: List recent novel sessions
      description: Returns compact session metadata for choosing which novel to resume.
      x-openai-isConsequential: false
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        "200":
          description: Recent sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionSummary"
  /v1/sessions/{session_id}:
    get:
      operationId: getSessionStatus
      summary: Resume a session
      description: Returns phase, versions, pending turn, public bootstrap review, and a compact current summary.
      x-openai-isConsequential: false
      parameters:
        - name: session_id
          in: path
          required: true
          description: Exact session ID returned by createSession or resumeSession.
          schema:
            type: string
            pattern: "^[a-zA-Z0-9_-]{1,100}$"
      responses:
        "200":
          description: Session status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionSummary"
  /v1/sessions/resume/{resume_code}:
    get:
      operationId: resumeSession
      summary: Resume a session by private code
      description: Resolves the private resume code returned at creation without enumerating other sessions.
      x-openai-isConsequential: false
      parameters:
        - name: resume_code
          in: path
          required: true
          description: Exact private resume code previously returned for this session.
          schema:
            type: string
            pattern: "^[A-Fa-f0-9]{8,12}$"
      responses:
        "200":
          description: Matching session status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionSummary"
  /v1/sessions/{session_id}/questionnaire:
    put:
      operationId: saveQuestionnaire
      summary: Save questionnaire answers
      description: >-
        Saves clarification or repairs a legacy empty session. If starter answers
        are already visible, send them from the conversation without asking the
        user to repeat. Identical retries are idempotent.
      x-openai-isConsequential: false
      parameters:
        - name: session_id
          in: path
          required: true
          description: Exact session ID returned by createSession or resumeSession.
          schema:
            type: string
            pattern: "^[a-zA-Z0-9_-]{1,100}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionnaireRequest"
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionSummary"
  /v1/sessions/{session_id}/bootstrap/parts:
    post:
      operationId: saveBootstrapPart
      summary: Save one generated bootstrap part
      description: >-
        Saves profile, lore, hidden canon, plot, current state, one character
        card, or the public review without echoing it. Set merge=true when
        adding invented missing fields or repairing a saved part so existing
        user-derived data is not erased.
      x-openai-isConsequential: false
      parameters:
        - name: session_id
          in: path
          required: true
          description: Exact session ID returned by createSession or resumeSession.
          schema:
            type: string
            pattern: "^[a-zA-Z0-9_-]{1,100}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BootstrapPartRequest"
      responses:
        "200":
          description: Small save receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BootstrapSaveResponse"
  /v1/sessions/{session_id}/bootstrap/validate:
    post:
      operationId: validateBootstrap
      summary: Validate the generated project
      description: >-
        Reports director_repairs separately from user_questions. The director
        must invent and save every director repair, normally with merge=true.
        Ask the user only when next_action is ask_user. It does not activate the
        story.
      x-openai-isConsequential: false
      parameters:
        - name: session_id
          in: path
          required: true
          description: Exact session ID returned by createSession or resumeSession.
          schema:
            type: string
            pattern: "^[a-zA-Z0-9_-]{1,100}$"
      responses:
        "200":
          description: Bootstrap validation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BootstrapValidation"
  /v1/sessions/{session_id}/bootstrap/confirm:
    post:
      operationId: confirmBootstrap
      summary: Activate a confirmed novel
      description: Promotes a valid draft to canonical per-session state only after explicit user confirmation.
      x-openai-isConsequential: false
      parameters:
        - name: session_id
          in: path
          required: true
          description: Exact session ID returned by createSession or resumeSession.
          schema:
            type: string
            pattern: "^[a-zA-Z0-9_-]{1,100}$"
      responses:
        "200":
          description: Active session status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionSummary"
  /v1/sessions/{session_id}/turns:
    post:
      operationId: prepareTurn
      summary: Freeze context for one turn
      description: >-
        Freezes complete context from this session, including required cards,
        knowledge, chronology, lore, and all hidden canon. It fails rather than
        omit context when the packet cannot fit. Identical retries return the
        same turn.
      x-openai-isConsequential: false
      parameters:
        - name: session_id
          in: path
          required: true
          description: Exact session ID returned by createSession or resumeSession.
          schema:
            type: string
            pattern: "^[a-zA-Z0-9_-]{1,100}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrepareTurnRequest"
      responses:
        "200":
          description: Prepared turn and first context chunk
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreparedTurn"
  /v1/sessions/{session_id}/turns/{turn_id}/chunks/{chunk_index}:
    get:
      operationId: getTurnChunk
      summary: Read another frozen context chunk
      description: >-
        Reads the exact packet created by prepareTurn. Read every chunk until
        has_more is false before writing prose. It never recomputes context
        from live state.
      x-openai-isConsequential: false
      parameters:
        - name: session_id
          in: path
          required: true
          description: Exact session ID returned by createSession or resumeSession.
          schema:
            type: string
            pattern: "^[a-zA-Z0-9_-]{1,100}$"
        - name: turn_id
          in: path
          required: true
          description: Exact frozen turn ID returned by prepareTurn.
          schema:
            type: string
            pattern: "^[a-zA-Z0-9_-]{1,100}$"
        - name: chunk_index
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
            maximum: 7
      responses:
        "200":
          description: Requested frozen chunk
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreparedTurn"
  /v1/sessions/{session_id}/turns/{turn_id}/commit:
    post:
      operationId: commitTurn
      summary: Persist a scene and its state changes
      description: >-
        Validates scene presentation, tenth-turn continuity audit requirements,
        typed state changes, and the frozen base version; then saves
        transactionally and returns an idempotent receipt. Technical and audit
        commits cannot advance story time.
      x-openai-isConsequential: false
      parameters:
        - name: session_id
          in: path
          required: true
          description: Exact session ID returned by createSession or resumeSession.
          schema:
            type: string
            pattern: "^[a-zA-Z0-9_-]{1,100}$"
        - name: turn_id
          in: path
          required: true
          description: Exact frozen turn ID returned by prepareTurn.
          schema:
            type: string
            pattern: "^[a-zA-Z0-9_-]{1,100}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitTurnRequest"
      responses:
        "200":
          description: Commit receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TurnReceipt"
  /v1/sessions/{session_id}/turns/{turn_id}/abort:
    post:
      operationId: abortTurn
      summary: Abort an uncommitted prepared turn
      description: Closes a pending turn without changing story state. A committed turn cannot be aborted.
      x-openai-isConsequential: false
      parameters:
        - name: session_id
          in: path
          required: true
          description: Exact session ID returned by createSession or resumeSession.
          schema:
            type: string
            pattern: "^[a-zA-Z0-9_-]{1,100}$"
        - name: turn_id
          in: path
          required: true
          description: Exact frozen turn ID returned by prepareTurn.
          schema:
            type: string
            pattern: "^[a-zA-Z0-9_-]{1,100}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  maxLength: 1000
              additionalProperties: false
      responses:
        "200":
          description: Abort status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  turn_id:
                    type: string
                required: [status, turn_id]
                additionalProperties: false
components:
  schemas:
    StartQuestionnaireResponse:
      type: object
      required: [version, questionnaire]
      properties:
        version:
          type: string
        questionnaire:
          type: string
          maxLength: 12000
          description: Show this Markdown text to the user exactly as returned.
      additionalProperties: false
    CreateSessionRequest:
      type: object
      required: [raw_answers]
      properties:
        title:
          type: [string, "null"]
          maxLength: 160
        raw_answers:
          type: string
          minLength: 1
          maxLength: 100000
          description: Preserve all visible starter answers verbatim; never summarize or request them again.
        normalized:
          type: [object, string]
          description: Every explicit user fact as small structured values; JSON text is accepted as fallback.
        unknown_fields:
          type: [array, string]
          description: Ordinary gaps the director will invent without asking the user.
          items:
            type: string
        contradictions:
          type: [array, string]
          description: Only unresolved material conflicts that genuinely require a user answer.
          items:
            type: string
      additionalProperties: false
    SessionSummary:
      type: object
      required:
        - session_id
        - resume_code
        - title
        - status
        - schema_version
        - state_version
        - turn_number
        - created_at
        - updated_at
        - questionnaire_entry_count
      properties:
        session_id:
          type: string
        resume_code:
          type: string
        title:
          type: string
        status:
          type: string
        schema_version:
          type: integer
        state_version:
          type: integer
        turn_number:
          type: integer
        created_at:
          type: string
        updated_at:
          type: string
        pending_turn_id:
          type: [string, "null"]
        bootstrap_missing:
          type: array
          items:
            type: string
        bootstrap_warnings:
          type: array
          items:
            type: string
        questionnaire_entry_count:
          type: integer
        last_questionnaire_entry_id:
          type: [string, "null"]
        review:
          type: [object, "null"]
          additionalProperties: true
        current_summary:
          type: [object, "null"]
          additionalProperties: true
      additionalProperties: false
    QuestionnaireRequest:
      type: object
      required: [phase, raw_answers]
      properties:
        phase:
          type: string
          enum: [initial, clarification]
        raw_answers:
          type: string
          minLength: 1
          maxLength: 100000
          description: Preserve the user's answer exactly; do not summarize it.
        normalized:
          type: [object, string]
          description: >-
            Put every explicit user fact into small structured values rather
            than a prose summary. Bootstrap validation checks that each value
            reaches generated state. JSON text is accepted as a recovery fallback.
        unknown_fields:
          type: [array, string]
          description: Ordinary gaps the director will invent and save without questioning the user.
          items:
            type: string
        contradictions:
          type: [array, string]
          description: Only unresolved material conflicts that genuinely require a user answer.
          items:
            type: string
      additionalProperties: false
    BootstrapPartRequest:
      type: object
      required: [part_type, content]
      description: One bootstrap document. Send content as a compact JSON object encoded as text.
      properties:
        part_type:
          type: string
          enum: [profile, lore, hidden_canon, plot, current, character, review]
        part_id:
          type: [string, "null"]
          maxLength: 80
        merge:
          type: boolean
          default: false
          description: Deep-merge this content into the saved part instead of replacing it.
        content:
          type: string
          maxLength: 60000
          description: >-
            A valid JSON object encoded as text. Example for profile:
            {"title":"My novel","genre":["romance"],"pov_id":"pov"}.
      additionalProperties: false
    BootstrapSaveResponse:
      type: object
      required: [status, part_type, size_chars, merged, warnings]
      properties:
        status:
          type: string
        part_type:
          type: string
        part_id:
          type: [string, "null"]
        size_chars:
          type: integer
        merged:
          type: boolean
        warnings:
          type: array
          items:
            type: string
      additionalProperties: false
    BootstrapValidation:
      type: object
      required:
        - ready
        - missing
        - errors
        - warnings
        - character_ids
        - director_repairs
        - user_questions
        - next_action
      properties:
        ready:
          type: boolean
        missing:
          type: array
          items:
            type: string
        errors:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        character_ids:
          type: array
          items:
            type: string
        director_repairs:
          type: array
          description: Missing or incomplete data the director must invent and save itself.
          items:
            type: string
        user_questions:
          type: array
          description: Only material contradictions that cannot be safely invented.
          items:
            type: string
        next_action:
          type: string
          enum: [show_review, repair_bootstrap, ask_user]
      additionalProperties: false
    PrepareTurnRequest:
      type: object
      properties:
        user_input:
          type: string
          maxLength: 30000
          default: ""
        mode:
          type: string
          enum: [play, technical, audit]
          default: play
        replace_pending:
          type: boolean
          default: false
      additionalProperties: false
    PreparedTurn:
      type: object
      required:
        - status
        - session_id
        - turn_id
        - mode
        - base_state_version
        - input_hash
        - chunk
        - has_more
        - total_chunks
        - audit_due
        - context_complete
        - included_sections
        - warnings
      properties:
        status:
          type: string
        session_id:
          type: string
        turn_id:
          type: string
        mode:
          type: string
        base_state_version:
          type: integer
        input_hash:
          type: string
        chunk:
          type: object
          properties:
            chunk_index:
              type: integer
            sections:
              type: array
              items:
                type: object
                additionalProperties: true
          required: [chunk_index, sections]
          additionalProperties: false
        has_more:
          type: boolean
        next_chunk_index:
          type: [integer, "null"]
        total_chunks:
          type: integer
          maximum: 8
        audit_due:
          type: boolean
          description: True when this commit must include a complete continuity audit.
        context_complete:
          type: boolean
          enum: [true]
          description: A prepared turn is returned only when no required context was omitted.
        included_sections:
          type: array
          description: Exact frozen packet section names delivered across all chunks.
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
      additionalProperties: false
    CharacterPatch:
      type: object
      required: [character_id, changes]
      properties:
        character_id:
          type: string
        changes:
          type: object
          additionalProperties: true
      additionalProperties: false
    NewCharacter:
      type: object
      required: [character_id, card]
      properties:
        character_id:
          type: string
        card:
          type: object
          additionalProperties: true
        starting_knowledge:
          type: array
          items:
            type: object
            additionalProperties: true
      additionalProperties: false
    KnowledgeEvent:
      type: object
      required: [character_id, fact, status, source]
      properties:
        character_id:
          type: string
        fact:
          type: string
        status:
          type: string
          enum:
            - fact
            - heard_fragment
            - incomplete_version
            - mistaken_version
            - suspicion
            - lie_believed
            - corrected
            - refuted
            - unknown
        source:
          type: string
        replaces_entry_id:
          type: [string, "null"]
      additionalProperties: false
    RelationshipEvent:
      type: object
      required: [from_character_id, to_character_id, metric, delta, reason]
      properties:
        from_character_id:
          type: string
        to_character_id:
          type: string
        metric:
          type: string
        delta:
          type: integer
          minimum: -25
          maximum: 25
        reason:
          type: string
      additionalProperties: false
    PlotlinePatch:
      type: object
      required: [plotline_id, changes]
      properties:
        plotline_id:
          type: string
        changes:
          type: object
          additionalProperties: true
      additionalProperties: false
    CommitTurnRequest:
      type: object
      properties:
        scene_text:
          type: string
          maxLength: 20000
          default: ""
        scene_summary:
          type: string
          maxLength: 5000
          default: ""
        current_patch:
          type: object
          additionalProperties: true
        time_advance_minutes:
          type: integer
          minimum: 0
          maximum: 525600
          default: 0
        character_patches:
          type: array
          items:
            $ref: "#/components/schemas/CharacterPatch"
        new_characters:
          type: array
          items:
            $ref: "#/components/schemas/NewCharacter"
        knowledge_events:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgeEvent"
        relationship_events:
          type: array
          items:
            $ref: "#/components/schemas/RelationshipEvent"
        plotline_patches:
          type: array
          items:
            $ref: "#/components/schemas/PlotlinePatch"
        chronology_event:
          type: [object, "null"]
          additionalProperties: true
        audit_updates:
          $ref: "#/components/schemas/AuditUpdates"
      additionalProperties: false
    AuditUpdates:
      type: object
      properties:
        continuity_checked:
          type: boolean
        chronology_checked:
          type: boolean
        checked_character_ids:
          type: array
          items:
            type: string
        issues:
          type: array
          items: {}
        repairs:
          type: array
          items: {}
      additionalProperties: true
    TurnReceipt:
      type: object
      required:
        - status
        - session_id
        - turn_id
        - mode
        - base_state_version
        - new_state_version
        - turn_number
        - changed_files
        - warnings
        - committed_at
      properties:
        status:
          type: string
        session_id:
          type: string
        turn_id:
          type: string
        mode:
          type: string
        base_state_version:
          type: integer
        new_state_version:
          type: integer
        turn_number:
          type: integer
        scene_number:
          type: [integer, "null"]
        changed_files:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        committed_at:
          type: string
      additionalProperties: false